moved messageBuilderFactory to a src/utils directory
